tensor_product: Fix shape of dual basis
We can't just naively index the dual of the factors and outer product
them because we need to split between index_shape and value_shape and
then produce a final dual with (*index_shapes, *value_shapes). So grab
some indices for each factor, product those, and then glue the new
indices together in the correct order.